pyproject.toml
[tool.poetry]
name = "google-meet-api"
version = "0.0.6"
description = "This is a tool to get IP and system information of a specific device"
authors = ["Subhomoy Roy Choudhury <subhomoyrchoudhury@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/subhomoy-roy-choudhury/Hack_IP"
repository = "https://github.com/subhomoy-roy-choudhury/Hack_IP"
keywords = ["python", "google-meet", "api"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "src/meet_api"}]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.86.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
flake8 = "^6.1.0"
tox = "^4.11.4"
poetry-dynamic-versioning = "^1.2.0"
pre-commit = { version = "^3.6.0", python = ">=3.9" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
